long

+++ src/main/java/com/dkbyte/project/repository/ConductorRepository.java
@@ -2,5 +2,5 @@ package com.dkbyte.project.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConductorRepository extends JpaRepository<ConductorRepository, Integer> {
+public interface ConductorRepository extends JpaRepository<ConductorRepository, Long> {
 }
