+++ src/main/java/com/dkbyte/project/repository/PersonaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonaRepository extends JpaRepository<PersonaRepository, Integer> {
+public interface PersonaRepository extends JpaRepository<PersonaRepository, Long> {
 }
